fix(security): stop exposing encrypted infrastructure secrets via server actions
getInfrastructuresIncludingDefault/getInfrastructureByName/
fetchAllInfrastructureConnectionsFromDb returned InfrastructureConnection
objects including infrastructureSecurity (encrypted header/proxy values,
init vectors, auth tags). As exports of a 'use server' module they were
HTTP-reachable server actions, and ManualAasInput referenced one from
client code, so any user could POST to it and retrieve the encrypted
secrets.

Move these secret-returning functions into a new plain server module
infrastructureData.ts (no 'use server'), importable only by server code.
Add a names-only getInfrastructureNamesAction for the UI and switch
ManualAasInput to it. Repoint server-side importers and their spec mocks.
getDefaultInfrastructure stays an action (env-only, no DB secrets).
