+++ src/lib/services/database/infrastructureDatabaseActions.ts
@@ -7,13 +7,12 @@ import { requireAdmin } from 'lib/util/securityHelpers/authGuard';
 import type { InfrastructureFormData } from 'app/[locale]/settings/_components/mnestix-infrastructure/InfrastructureTypes';
 import { envs } from 'lib/env/MnestixEnv';
 import { ConnectionTypeEnum, getTypeAction } from 'lib/services/database/ConnectionTypeEnum';
+import { InfrastructureConnection, RepositoryWithInfrastructure } from 'lib/services/database/InfrastructureMappedTypes';
 import {
-    InfrastructureConnection,
-    InfrastructureWithRelations,
-    RepositoryWithInfrastructure,
-} from 'lib/services/database/InfrastructureMappedTypes';
-
-const DEFAULT_INFRASTRUCTURE_NAME = 'Default Infrastructure';
+    buildDefaultInfrastructure,
+    DEFAULT_INFRASTRUCTURE_NAME,
+    getInfrastructuresIncludingDefault,
+} from 'lib/services/database/infrastructureData';
 
 export async function getDefaultInfrastructureName() {
     return DEFAULT_INFRASTRUCTURE_NAME;
@@ -30,48 +29,20 @@ export async function getConnectionDataByTypeAction(type: ConnectionType): Promi
     return prismaConnector.getConnectionDataByTypeAction(type);
 }
 
-export async function fetchAllInfrastructureConnectionsFromDb(): Promise<InfrastructureConnection[]> {
-    const connector = PrismaConnector.create();
-    const infrastructures = await connector.getInfrastructures();
-
-    if (!infrastructures) return [];
-
-    return infrastructures.map((infra) => infrastructureMapper(infra));
-}
-
 export async function getDefaultInfrastructure(): Promise<InfrastructureConnection> {
-    return {
-        name: DEFAULT_INFRASTRUCTURE_NAME,
-        discoveryUrls: envs.DISCOVERY_API_URL ? [envs.DISCOVERY_API_URL] : [],
-        aasRegistryUrls: envs.REGISTRY_API_URL ? [envs.REGISTRY_API_URL] : [],
-        aasRepositoryUrls: envs.AAS_REPO_API_URL ? [envs.AAS_REPO_API_URL] : [],
-        submodelRepositoryUrls: envs.SUBMODEL_REPO_API_URL ? [envs.SUBMODEL_REPO_API_URL] : [],
-        submodelRegistryUrls: envs.SUBMODEL_REGISTRY_API_URL ? [envs.SUBMODEL_REGISTRY_API_URL] : [],
-        conceptDescriptionRepositoryUrls: envs.CONCEPT_DESCRIPTION_REPO_API_URL
-            ? [envs.CONCEPT_DESCRIPTION_REPO_API_URL]
-            : [],
-        serializationEndpointUrls: envs.SERIALIZATION_API_URL ? [envs.SERIALIZATION_API_URL] : [],
-        isDefault: true,
-    };
+    // Env-derived only; carries no database security secrets, so it is safe to expose as an action.
+    return buildDefaultInfrastructure();
 }
 
-export async function getInfrastructuresIncludingDefault() {
-    // build default infrastructure from envs
-    const defaultInfrastructure = await getDefaultInfrastructure();
-
-    // get from database as flat connection list
-    const infrastructures = await fetchAllInfrastructureConnectionsFromDb();
-
-    return [defaultInfrastructure, ...infrastructures];
-}
-
-export async function getInfrastructureByName(name: string): Promise<InfrastructureConnection | undefined> {
+/**
+ * UI-facing action returning only infrastructure names. Deliberately does NOT
+ * return the full `InfrastructureConnection` shape, which includes encrypted
+ * security configuration. Use this from client components instead of the
+ * server-internal `getInfrastructuresIncludingDefault`.
+ */
+export async function getInfrastructureNamesAction(): Promise<string[]> {
     const infrastructures = await getInfrastructuresIncludingDefault();
-    const found_infrastructure = infrastructures.find((infra) => infra.name === name);
-    if (!found_infrastructure) {
-        return undefined;
-    }
-    return found_infrastructure;
+    return infrastructures.map((infra) => infra.name);
 }
 
 export async function getAasRepositoriesIncludingDefault() {
@@ -137,50 +108,4 @@ export async function deleteInfrastructureAction(infrastructureId: string): Prom
     await requireAdmin();
     const connector = PrismaConnector.create();
     await connector.deleteInfrastructureAction(infrastructureId);
-}
-
-function infrastructureMapper(infra: InfrastructureWithRelations): InfrastructureConnection {
-    return {
-        name: infra.name,
-        discoveryUrls: infra.connections.flatMap((conn) =>
-            conn.types.filter((t) => t.type.typeName === 'DISCOVERY_SERVICE').map(() => conn.url),
-        ),
-        aasRegistryUrls: infra.connections.flatMap((conn) =>
-            conn.types.filter((t) => t.type.typeName === 'AAS_REGISTRY').map(() => conn.url),
-        ),
-        aasRepositoryUrls: infra.connections.flatMap((conn) =>
-            conn.types.filter((t) => t.type.typeName === 'AAS_REPOSITORY').map(() => conn.url),
-        ),
-        submodelRepositoryUrls: infra.connections.flatMap((conn) =>
-            conn.types.filter((t) => t.type.typeName === 'SUBMODEL_REPOSITORY').map(() => conn.url),
-        ),
-        submodelRegistryUrls: infra.connections.flatMap((conn) =>
-            conn.types.filter((t) => t.type.typeName === 'SUBMODEL_REGISTRY').map(() => conn.url),
-        ),
-        conceptDescriptionRepositoryUrls: infra.connections.flatMap((conn) =>
-            conn.types.filter((t) => t.type.typeName === 'CONCEPT_DESCRIPTION').map(() => conn.url),
-        ),
-        serializationEndpointUrls: infra.connections.flatMap((conn) =>
-            conn.types.filter((t) => t.type.typeName === 'SERIALIZATION_ENDPOINT').map(() => conn.url),
-        ),
-        isDefault: false,
-        infrastructureSecurity: {
-            securityType: infra.securityType.typeName,
-            securityHeader: infra.securitySettingsHeaders
-                ? {
-                      name: infra.securitySettingsHeaders.headerName,
-                      value: infra.securitySettingsHeaders.headerValue,
-                      initVector: infra.securitySettingsHeaders.initVector,
-                      authTag: infra.securitySettingsHeaders.authTag,
-                  }
-                : undefined,
-            securityProxy: infra.securitySettingsProxies
-                ? {
-                      value: infra.securitySettingsProxies.headerValue,
-                      initVector: infra.securitySettingsProxies.initVector,
-                      authTag: infra.securitySettingsProxies.authTag,
-                  }
-                : undefined,
-        },
-    };
 }
